restructured product
