+++ src/front/js/pages/home.js
@@ -1,24 +1,27 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext, useEffect, Component } from "react";
 import { Context } from "../store/appContext";
 import ScriptTag from "react-script-tag";
 import { Carta } from "../component/Cards";
+import { Container } from "react-bootstrap";
+import { Row } from "react-bootstrap";
+import { Col } from "react-bootstrap";
+import { Button } from "react-bootstrap";
+import { Aboutus } from "../component/AboutUs";
+import { Contact } from "../component/Contact";
 
 export const Home = () => {
 	return (
 		<div>
 			<div className="jumbotron jumbotron-fluid">
-				<div className="container">
-					<h1 className="display-2">
-						A Brand New Way <br />
-						to Access Your <br />
-						Favorite Cocktail <br /> Recipes
-					</h1>
-					<a className="btn btn-primary btn-lg" href="#" role="button">
-						Get Started
-					</a>
-				</div>
+				<Container>
+					<Row>
+						<Col>
+							<h1 className="display-2">Insert Text</h1>
+							<Button>Show More</Button>
+						</Col>
+					</Row>
+				</Container>
 			</div>
-
 			<br />
 			<div
 				className="cards d-flex justify-content-center"
@@ -31,6 +34,12 @@ export const Home = () => {
 				<Carta />
 				<ScriptTag src="/workspace/Final-Project-/node_modules/aos/dist/aos.js" />
 			</div>
+			<br />
+			<br />
+			<Aboutus />
+			<br />
+			<br />
+			<Contact />
 		</div>
 	);
 };
